Add UID, prefer printf, and try HOSTNAME
If, for whatever reason, the `HOSTNAME` variable is empty or not
assigned, then it'll try your approach of the `hostname` command. It's
just a bit more efficient and cleaner.

The `UID` variable could be switched to `EUID`, if preferred.

Cleaned up the `read` usage, since `-e` will do as desired.

Used `printf` for a builtin solution to gathering the time and date. I
have used the same output as `date`, in-case you were specifically
wanted it. The use of `-v` sets and assigns the shell variable.