Add noteid fallback and improve usage output
The `${0##*/}` has the same effect as using `basename`, but is more
efficient (one fewer process), provided you're not using the Bourne
Shell, as is _not_ the case here.

I added `this` to your `where` part of the usage information, so the
user doesn't incorrectly assume there's further information.

I changed the help message if statement to case, for cleaner code and
I believe it'd make further changes a little easier.