Overhaul for it to be cleaner and smarter
Now your script is more robust and presentable. I think you'll like it.

If you're not keen on things like the double-quotes and/or the author
part (like you typically see when someone is quote) then an option can
be added to remove it, to make it easily parseable.

The help output got an overhaul. Inspired from my personal preference
on my own programs, although that itself is more like man pages and
often is similar to traditional help output. Hope it's okay.